Manage global state and navigation in BaseActivity
Keep global state in a private companion object of BaseActivity. For
now, only use a HashMap to store Models using its host activity as key.
BaseActivity will update this HashMap whenever a controller requests to
go to a different scene.

Add new method to IHostActivity: `goToScene` which receives a params
object which BaseActivity will use to create the new model along with
the new activity. Since BaseActivity now creates the models and keeps
references in the global state, the `initController` function now
receives the model as parameter.