Write signal tests with room
New classes TestDatabase and ResetDao to be used only in instrumentation tests.
New test LocalPersistentCommunicationTest creates one user in Room and
another in memory and makes them exchange messages.
Delete and insert session in SignalStoreCriptext's store session to
avoid errors when signal wants to update a session.
Add `complete()` method to IncompleteAccount that returns an Account
object.
InMemoryUser now subclasses new class TestUser. this makes it easier to
create instances of users in memory and in database