New class EmailInsertionSetup
EmailInsertionSetup is an object that exposes a single function `exec`
which is the simplest abstraction over inserting an email. This function
should be called inside a database transactions so that a new email is
inserted along with all the relations to contacts and labels.

UpdateMailboxWorker has been refactored to use this object. The result
now contains a nullable list of threads. It has to be nullable because
most of the time there wont be any updates.

Add a test for pull down to refresh in MailboxSceneController.