Create instrumentation test for SaveEmailWorker
Create class EmailMetadata.DBColumns to use only the subset of
EmailMetadata that gets inserted to DB.
Add functions to DateUtils to convert timestamps to date back and forth
with the server's format.
Delete the previously created draft, if any, in the SaveEmailWorker
transaction.
Delete emails, contacts in ResetDao.
Create instrumentation test for SaveEmailWorker, asserting that
everything gets inserted to the DB correctly