Create UploadAttachmentWorker
Create ChunkFileReader class to abstract an algorithm for reading files
by chunks of fixed size.
Add a new method to HttpClient to post multipart form data so that
binary data is encoded correctly in the http request.
Create FileServiceAPIClient to interact with the File Service via HTTP.
move ListUtils to an appropriate package.